Move BooksForm and BookList to containers folder
